+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/extensions/SnapshotPreviewsExtension.kt
@@ -1,19 +1,18 @@
-package io.sentry.android.gradle.snapshot
+package io.sentry.android.gradle.extensions
 
+import javax.inject.Inject
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
 import org.jetbrains.annotations.ApiStatus
 
-/**
- * Experimental extension for configuring Compose @Preview snapshot testing. This API is subject to
- * change and will eventually be merged into the main `sentry` extension.
- */
 @ApiStatus.Experimental
-abstract class SentrySnapshotExtension(objects: ObjectFactory) {
+open class SnapshotPreviewsExtension @Inject constructor(objects: ObjectFactory) {
+
+  val generateTests: Property<Boolean> = objects.property(Boolean::class.java).convention(true)
 
   val includePrivatePreviews: Property<Boolean> =
-    objects.property(Boolean::class.java).convention(false)
+    objects.property(Boolean::class.java).convention(true)
 
   val packageTrees: ListProperty<String> =
     objects.listProperty(String::class.java).convention(emptyList())