+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/instrumentation/fakes/TestSpanAddingParameters.kt
@@ -6,15 +6,19 @@ import io.sentry.android.gradle.instrumentation.SpanAddingClassVisitorFactory
 import io.sentry.android.gradle.instrumentation.logcat.LogcatLevel
 import io.sentry.android.gradle.services.SentryModulesService
 import java.io.File
+import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.internal.provider.DefaultProperty
 import org.gradle.api.internal.provider.PropertyHost
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.SetProperty
+import org.gradle.testfixtures.ProjectBuilder
 
 class TestSpanAddingParameters(
   private val debugOutput: Boolean = true,
   private val inMemoryDir: File,
 ) : SpanAddingClassVisitorFactory.SpanAddingParameters {
+
+  private val objects = ProjectBuilder.builder().build().objects
 
   override val invalidate: Property<Long>
     get() = DefaultProperty(PropertyHost.NO_OP, Long::class.java).convention(0L)
@@ -30,8 +34,8 @@ class TestSpanAddingParameters(
   override val sentryModulesService: Property<SentryModulesService>
     get() = TODO()
 
-  override val tmpDir: Property<File>
-    get() = DefaultProperty<File>(PropertyHost.NO_OP, File::class.java).convention(inMemoryDir)
+  override val tmpDir: DirectoryProperty
+    get() = objects.directoryProperty().also { it.set(inMemoryDir) }
 
   override var _instrumentable: ClassInstrumentable? = null