+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/telemetry/SentryTelemetryServiceTest.kt
@@ -17,12 +17,12 @@ class SentryTelemetryServiceTest {
     val project = ProjectBuilder.builder().withProjectDir(testProjectDir.root).build()
 
     val cliPath =
-      SentryCliProvider.getCliResourcesExtractionPath(project.layout.buildDirectory.asFile.get())
+      SentryCliProvider.getCliResourcesExtractionPath(project.layout.buildDirectory).get().asFile
 
     val infoOutput =
       project.providers
         .of(SentryCliInfoValueSource::class.java) { cliVS ->
-          cliVS.parameters.buildDirectory.set(project.buildDir)
+          cliVS.parameters.buildDirectory.set(project.layout.buildDirectory)
           cliVS.parameters.cliExecutable.set(cliPath.absolutePath)
           // sets an empty/invalid auth token
           cliVS.parameters.authToken.set("")